feat: 🎸 Add FieldConnector lastRemoteValue child prop (#40)
* feat: 🎸 Add FieldConnector lastRemoteValue child prop

The FieldConnector now informs about the last value that was set via an
incoming remote change via `lastRemoteValue`. It no longer passes
`lastSetValue` as it is conceptually hard to guess that its value will
be either the last value set via `setValue()` immediately after calling
it or the last incoming remote change value.