fix: 🐛 rt url modal auto focus not working due to editor focus
The editor needs to be blured before opening the url dialog for the
dialog's input auto focus to work. Otherwise, focus stays with Slate,
even while the modal is open, until the user clicks on a modal input.
This also means if the user were to write while the modal is open before
clicking on an input, the text would be added in the background to the
editor, potentially causing other bugs when inserting the link.