feat: 🎸 refined editor focusing when using toolbar actions (#350)
When the user clicks on a Rich Text editor toolbar action while the
editor is not focused, this should ideally result in focusing the editor
and applying the clicked action.

This handling has been implemented for some actions and it has been
breaking in some cases. This change introduces consistent handling and
dealing with Slate related limitations.

The main limitaton would be after initializing Rich Text, before the
user has manually focused the rich text field, we can not set the focus
programmatically as the Slate internal value seems to lack some
properties that can result in a crash at this point. We circumvent bugs
by preventing any action when clicking on a toolbar action. Instead, the
user has to focus the input first. This is pretty much consistent with
the previous behavior but implemented in a way that is able to prevent
any crashes. Ideally we would improve the behaivor once we figure out
how to prevent Slate from having these missing properties or after
upgrading to a newer Slate 0.5x version.