fix: pass already known entry link id to onAction
Instead of passing the fetched entry's `sys.id` to `onAction` events,
simply pass the already known entry ID prop. We know the ID as it's part
of the actual link.
This has the big advantage that even when the entry can't be loaded, or
if an `onAction` is called before it loaded, we can still send teh entry
ID.