fix: do not focus RTE when another slide is open
In the rich text editor, when embedding an entry or asset block, ensure
that we take into account the entity selector option of `withCreate: true`
which can cause the slide-in editor to open (web app).
When the slide-in editor opens, then we only want to re-focus the rich
text editor after the slide-in editor (and any other slides opened on
top of that) closes.