fix: 🐛 ensure watchCurrentSlide handles noop slide event
In the web app, when clicking on a reference that references the
current entry, the web app's `sdk` still fires a slide-in event with
`{ oldSlideLevel: 0, newSlideLevel: 0 }`. This should not in turn fire
any `watchCurrentSlide().onActive(callback)` because the slide already
was active.
There's no particular downstream bug that this is fixing as it does
not really matter where we currently use the helper.