fix: 🐛 do not focus RTE on inline embed + create entry
Same as db07c18 but for inline embeds. When we create a new entry for an
inline embed in the web app, the slide-in editor opens. In this case we
want to focus the slided in entry editor's first field, not the rich
text editor in the background tab. Only after the slide closes again, we
will focus the rich text editor again.