fix: 🐛 do not focus RTE on inline embed + create entry (#1354)
* fix: 🐛 do not focus RTE on inline embed + create entry

Same as db07c18 but for inline embeds. When we create a new entry for an
inline embed in the web app, the slide-in editor opens. In this case, we
want to focus the slided in entry editor's first field, not the rich
text editor in the background tab. Only after the slide closes again, we
will focus the rich text editor again.

* fix: 🐛 ensure watchCurrentSlide handles noop slide event

In the web app, when clicking on a reference that references the
current entry, the web app's `sdk` still fires a slide-in event with
`{ oldSlideLevel: 0, newSlideLevel: 0 }`. This should not in turn fire
any `watchCurrentSlide().onActive(callback)` because the slide already
was active.
There's no particular downstream bug that this is fixing as it does
not really matter where we currently use the helper.